torchmods/requirements.txt
torch==1.7.0+cu110
matplotlib==3.1.1
nvidia_ml_py3==7.352.0
paramiko==2.7.1
opencv_python==4.1.0.25
requests==2.22.0
numpy==1.16.5
torchvision==0.8.1+cu110
beautifulsoup4==4.9.3
imageio==2.9.0
Pillow==8.1.0
pynvml==8.0.4
